+++ test/js/strings.test.js
@@ -49,8 +49,8 @@ describe("strings", () => {
       test("interpolation within double quotes stay", () =>
         expect(`"#{"\\n"}"`).toMatchFormat());
 
-      test("escaped double quotes are unquoted", () =>
-        expect("'abc \\\"def\\\" ghi'").toChangeFormat("'abc \"def\" ghi'"));
+      test("escaped double quotes are not unquoted", () =>
+        expect("'abc \\\"def\\\" ghi'").toMatchFormat());
     });
   });