+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Changed
+
+- [@kddeisz] - Do not unescape double quotes in a single quote string.
+
 ## [1.0.0-rc2] - 2020-12-10
 
 ### Changed